refactor: move `Combobox` component from `embed` to `components` package
