+++ packages/components/src/components/index.js
@@ -1,0 +1,3 @@
+import ComboboxComponents from './Combobox';
+
+export const Combobox = ComboboxComponents;
